Rewrite start of drag-n-drop.
Because now offset/zoom are float, logic similar to how it worked in
Gtk2::ImageView doesn't always work anymore.

Instead, do it on motion if the mouse movement didn't move the image
which was already at the edge. What exactly should happen on drag-n-drop
is configurable in user code via new dnd-start signal.